Create Control Statements - Treasure Finder
Input consists of three integers. The first input corresponds to the number of the first box. The second input corresponds to the number of the second box. The third input corresponds to the number of the third box.

Sample Input :

2
4
6

Sample Output :

The treasure is in the box which has the number 4
The code to open the box is 2